+++ config/healthchecks.php
@@ -1,0 +1,18 @@
+<?php
+
+return [
+
+    /*
+     * The URL endpoint of healthchecks.io
+     */
+    'url' => 'https://hc-ping.com/',
+
+    /*
+     * Your registered jobs
+     */
+    'jobs' => [
+        'my-first-check' => [
+            'uuid' => 'c2c0be0a-94fa-4128-aa8a-cd55889cdb29',
+        ],
+    ],
+];