+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `laravel-healthchecks` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 1.0.0 - 2022-03-09
 
-- initial release
+- Initial release
+- Add `->success()`, `->failure()`, `->start()` and `->exitWithStatus($status)` methods